requirements.txt
pylinac==3.38.0
numpy==2.3.5
pillow==12.0.0
pydicom==2.4.4
